New Notes Task Diagram
This diagram is depicting the situation where the user creates a new note on the page https://studies.cs.helsinki.fi/exampleapp/notes by writing something into the text field and clicking the Save button. showing the processes of event on the site from the browser to the server, and vice reponses.