SPA Note Entary Diagram
This Diagram shows the process of event that happens between the browser and the server, were the user types a note and hits “save.” Immediately, JavaScript running in the browser springs into action. This code doesn’t reload the page; instead, it silently packages the note content (e.g., "The content of the new note") into a neat data object.